DataCollection/distributed/Perlmutter_LGPU_QAOA/requirements.txt
aiosignal==1.3.1
appdirs==1.4.4
attrs==23.2.0
autograd==1.6.2
autoray==0.6.8
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.3
contourpy==1.2.0
custatevec-cu12==1.5.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
future==0.18.3
idna==3.6
importlib-resources==6.1.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
matplotlib==3.8.3
msgpack==1.0.7
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
packaging==23.2
-e git+ssh://git@github.com/PennyLaneAI/pennylane.git@f7bba2b8a21319309b21250073492573f5ee967a#egg=PennyLane
PennyLane-Lightning==0.35.0.dev15
-e git+ssh://git@github.com/PennyLaneAI/pennylane-lightning.git@102d848cc4232afb181eed33c753efd4093467fb#egg=PennyLane_Lightning_GPU
pillow==10.2.0
protobuf==4.25.3
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
ray==2.9.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
rustworkx==0.14.0
scipy==1.12.0
semantic-version==2.10.0
six==1.16.0
toml==0.10.2
typing_extensions==4.9.0
urllib3==2.2.0
zipp==3.17.0
